pyproject.toml
[project]
name = "db-migration-tracking-alembic"
version = "0.1.0"
description = ""
authors = [
    {name = "AdityaZala3919",email = "117515629+AdityaZala3919@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "pathlib (>=1.0.1,<2.0.0)",
    "fastapi (>=0.135.1,<0.136.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "logging (>=0.4.9.6,<0.5.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)"
]

[tool.poetry]
packages = [{include = "db_migration_tracking_alembic", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
